Add single quotes around configurable string values in O365 (#25215)
Values passed in by users that are expected to be strings should be single-quoted.

Also, this fixes the `tojson` function to not escape &, <, and > to to \u0026, \u003c, and \u003e. This
caused problems if the value is an api keys or password that contained one of those characters.

Fixes #25058

(cherry picked from commit eed1cbb2dae431b15a37738f012706a6e564ae50)